Change order of conditionals to avoid broadcasting
The xarray.where() function used to return a new DataArray with
different dims than the original one. One way to prevent this was to
broadcast the coordinates arrays, although this consumed unnecessary
memory. By changing the order of the conditionals when calling the
xarray.where() function, the broadcasting is not needed saving memory.