Improve generation of temperature distributions (#5)
Create coordinates.py with new grid_coordinates and initialize_arrays
functions to build grid coordinates and a constant array from this
coordinates. Remove the empty_temperature_grid function because it's no
longer needed due to the new coordinates.py functions. Fix how the
temperature distribution of lithosphere and asthenosphere is created:
use the where method and create the asthenosphere temperature following
Blom (2016) (no dependency on lid_depth and use potential asthenosphere
temperature on surface). Allow to pass the LID depth as a constant value
or as variable profile (2D) or surface (3D). Update how the temperature
for a subducting slab is created: override temperature distribution on
the points inside the slab, and specify the slab through slope,
thickness, and horizontal boundaries. Also allow to change direction of
the subduction for 3D modelling. Update temperature examples for new
code.